requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.1.3
openai==2.7.2
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.5.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
